Final Version
This is the final contribution in this part of the project from my end, as the backtesting results are fairly affirmative.
I am getting expected returns out of this strategy as were desired by me. Obviously, as there is no one-fit-all strategy, this works well
only on the tickers mentioned in tickers.txt file among the NIFTY100 stocks.